+++ ff-test/test/Config.hs
@@ -1,39 +1,47 @@
+{-# LANGUAGE BlockArguments #-}
 {-# LANGUAGE NamedFieldPuns #-}
 {-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE TemplateHaskell #-}
 
 module Config (configTests) where
 
-import           Hedgehog (Property, evalIO, forAll, property, (===))
-import           System.Environment (setEnv)
-import           System.IO.Temp (withSystemTempDirectory)
-import           Test.Tasty (TestTree)
-import           Test.Tasty.Hedgehog (testProperty)
-import           Test.Tasty.TH (testGroupGenerator)
+import Hedgehog (Property, evalIO, forAll, property, withTests, (===))
+import System.Directory (withCurrentDirectory)
+import System.Environment (setEnv)
+import System.IO.Temp (withSystemTempDirectory)
+import Test.Tasty (TestTree)
+import Test.Tasty.Hedgehog (testProperty)
+import Test.Tasty.TH (testGroupGenerator)
 
-import           FF.Config (emptyConfig, loadConfig, saveConfig)
+import FF.Config (emptyConfig, loadConfig, saveConfig)
 
-import qualified Gen
+import Gen qualified
 
 configTests :: TestTree
-configTests = $(testGroupGenerator)
+configTests = $testGroupGenerator
 
 prop_loadNoConfig :: Property
-prop_loadNoConfig = property $ do
-    config <- evalIO $ withTempHome loadConfig
-    emptyConfig === config
+prop_loadNoConfig =
+    withTests 1 $
+        property do
+            config <-
+                evalIO . withCurrentDirectory "/" $ withTempHome loadConfig
+            emptyConfig === config
 
 prop_loadConfig :: Property
-prop_loadConfig = property $ do
-    config <- forAll Gen.config
-    loadedconf <-
-        evalIO $ withTempHome $ do
-            saveConfig config
-            loadConfig
-    config === loadedconf
+prop_loadConfig =
+    property do
+        config <- forAll Gen.config
+        loadedconf <-
+            evalIO
+                . withCurrentDirectory "/"
+                $ withTempHome do
+                    saveConfig config
+                    loadConfig
+        config === loadedconf
 
 withTempHome :: IO a -> IO a
 withTempHome action =
-    withSystemTempDirectory "ff-test.home" $ \tempHome -> do
+    withSystemTempDirectory "ff-test.home" \tempHome -> do
         setEnv "HOME" tempHome
         action